Add separate file for `ExtendTci` command.
